cleanup dependencies, remove remaining boilerplate code
