+++ app/lib/updater.js
@@ -45,6 +45,8 @@ Updater.prototype.check = function() {
       return self.emit('error', err || new Error('Failed to check updates'));
     }
 
+    console.log(body)
+
     try {
       self._validateResponse(body);
     } catch (err) {
